--- plugin.yml
@@ -1,9 +1,0 @@
-name: SurvivalGamesBGHD
-main: me.ItsJasonn.SurvivalGames.Main.Core
-version: 0.2
-author: BGHDDevelopment
-commands:
- Hub:
-  description: Teleport to the main hub
- SG:
-  description: Show a Survival Games help page
