--- src/me/ItsJasonn/SurvivalGames/Main/Plugin.java
@@ -1,14 +1,0 @@
-package me.ItsJasonn.SurvivalGames.Main;
-
-public class Plugin {
-	private static Core core;
-
-	@SuppressWarnings("static-access")
-	public Plugin(Core core) {
-		this.core = core;
-	}
-	
-	public static Core getCore() {
-		return core;
-	}
-}
