--- src/me/ItsJasonn/SurvivalGames/Utils/Server.java
@@ -1,30 +1,0 @@
-package me.ItsJasonn.SurvivalGames.Utils;
-
-import java.io.ByteArrayOutputStream;
-import java.io.DataOutputStream;
-import me.ItsJasonn.SurvivalGames.Main.Plugin;
-import org.bukkit.entity.Player;
-
-public class Server {
-	private String server = "";
-
-	public Server(String server) {
-		this.server = server;
-	}
-
-	public void sendPlayerToServer(Player player) {
-		ByteArrayOutputStream b = new ByteArrayOutputStream();
-		DataOutputStream out = new DataOutputStream(b);
-		try {
-			out.writeUTF("Connect");
-			out.writeUTF(this.server);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		player.sendPluginMessage(Plugin.getCore(), "BungeeCord", b.toByteArray());
-	}
-
-	public String getName() {
-		return this.server;
-	}
-}
