--- src/org/jnbt/EndTag.java
@@ -1,65 +1,0 @@
-/*    */ package org.jnbt;
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ public final class EndTag
-/*    */   extends Tag
-/*    */ {
-/*    */   public EndTag()
-/*    */   {
-/* 47 */     super("");
-/*    */   }
-/*    */   
-/*    */   public Object getValue()
-/*    */   {
-/* 52 */     return null;
-/*    */   }
-/*    */   
-/*    */   public String toString()
-/*    */   {
-/* 57 */     return "TAG_End";
-/*    */   }
-/*    */ }
-
-
-/* Location:              C:\Users\JJ\Downloads\Minecraft\Plugins\Dungeons.jar!\org\jnbt\EndTag.class
- * Java compiler version: 7 (51.0)
- * JD-Core Version:       0.7.1
- */
