--- src/org/jnbt/LongTag.java
@@ -1,78 +1,0 @@
-/*    */ package org.jnbt;
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ public final class LongTag
-/*    */   extends Tag
-/*    */ {
-/*    */   private final long value;
-/*    */   
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */ 
-/*    */   public LongTag(String name, long value)
-/*    */   {
-/* 54 */     super(name);
-/* 55 */     this.value = value;
-/*    */   }
-/*    */   
-/*    */   public Long getValue()
-/*    */   {
-/* 60 */     return Long.valueOf(this.value);
-/*    */   }
-/*    */   
-/*    */   public String toString()
-/*    */   {
-/* 65 */     String name = getName();
-/* 66 */     String append = "";
-/* 67 */     if ((name != null) && (!name.equals(""))) {
-/* 68 */       append = "(\"" + getName() + "\")";
-/*    */     }
-/* 70 */     return "TAG_Long" + append + ": " + this.value;
-/*    */   }
-/*    */ }
-
-
-/* Location:              C:\Users\JJ\Downloads\Minecraft\Plugins\Dungeons.jar!\org\jnbt\LongTag.class
- * Java compiler version: 7 (51.0)
- * JD-Core Version:       0.7.1
- */
